Move test page class into its own folder
We will be adding the new test page class to this folder next, it’ll
keep things more organised
